PNG issues fixed 1. Attendance not reflecting (random issue). 2. Return Screen - Alignment issue.
IVYCPG-4454
Star filter - not working

+++ ivycpg/src/main/java/com/ivy/ui/retailer/viewretailers/data/RetailerDataManagerImpl.java
@@ -94,7 +94,7 @@ public class RetailerDataManagerImpl implements RetailerDataManager {
                         while (c.moveToNext()) {
                             dateWisePlanBO = new DateWisePlanBo();
 
-                            dateWisePlanBO.setPlanId(c.getInt(0));
+                            dateWisePlanBO.setPlanId(c.getLong(0));
                             dateWisePlanBO.setDistributorId(c.getInt(1));
                             dateWisePlanBO.setUserId(c.getInt(2));
                              dateWisePlanBO.setDate(c.getString(3));
@@ -162,7 +162,7 @@ public class RetailerDataManagerImpl implements RetailerDataManager {
                         while (c.moveToNext()) {
                             dateWisePlanBO = new DateWisePlanBo();
 
-                            dateWisePlanBO.setPlanId(c.getInt(0));
+                            dateWisePlanBO.setPlanId(c.getLong(0));
                             dateWisePlanBO.setDistributorId(c.getInt(1));
                             dateWisePlanBO.setUserId(c.getInt(2));
                             dateWisePlanBO.setDate(c.getString(3));
